migrate DOM utils to new package
