+++ web/modules/components/src/index.ts
@@ -14,15 +14,10 @@ export { default as Toggle } from "./Toggle.svelte"
 export { default as Wikitext } from "./Wikitext.svelte"
 
 export * from "./lib/animation"
-export * from "./lib/controls"
-export * from "./lib/focus"
-export * from "./lib/gesture"
 export * from "./lib/media"
 export * from "./lib/modal"
 export * from "./lib/popper"
 export * from "./lib/tippy"
 export * from "./lib/portal"
-export * from "./lib/swipe"
-export * from "./lib/useragent"
 
 export { toast } from "./lib/toasts"
