+++ web/modules/dom/README.md
@@ -1,0 +1,1 @@
+# @wikijump/dom
