+++ web/modules/dom/src/index.ts
@@ -1,0 +1,13 @@
+export * from "./base-button"
+export * from "./base-tooltip-button"
+export * from "./custom-elements"
+export * from "./focus"
+export * from "./gesture"
+export * from "./hover"
+export * from "./key-handling"
+export * from "./observe"
+export * from "./scrolling"
+export * from "./svelte-adapters"
+export * from "./swipe"
+export * from "./useragent"
+export * from "./validation"
