+++ web/modules/dom/tsconfig.json
@@ -1,0 +1,7 @@
+{
+  "compilerOptions": {
+    "composite": true
+  },
+  "extends": "../../tsconfig.json",
+  "include": ["src/**/*", "tests/**/*", "../_types/**/*"]
+}
